fix: 회원 탈퇴 이동
