+++ server/api/src/main/kotlin/com/kw/api/domain/member/service/MemberService.kt
@@ -20,9 +20,7 @@ class MemberService(val memberRepository: MemberRepository) {
         return MemberInfoResponse.from(member)
     }
 
-    fun withdrawMember(member: Member) {
-        member.withdrawMember()
-    }
+
 
     private fun isNicknameUnique(nickname: String) {
         if(!memberRepository.existsByNickname(nickname)){
